+++ lib/crossbeams/layout/link.rb
@@ -4,7 +4,9 @@ module Crossbeams
   module Layout
     # A link renderer - for rendering a link outside a form.
     class Link
-      include PageNode
+      extend MethodBuilder
+
+      node_adders :csrf
       attr_reader :text, :url, :style, :behaviour, :css_class, :id, :visible, :text_size
 
       def initialize(options) # rubocop:disable Metrics/AbcSize